ca-api-backend/requirements.txt
fastapi>=0.116.0
fastapi[standard]
uvicorn[standard]
google-auth
aiohttp
pydantic
requests
fastmcp>=2.10.2
asyncio